# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marvol Facility Cleaning Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /issues:
    get:
      operationId: listIssues
      tags: [issues]
      summary: List all issues
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
          description: Start date for range filter (YYYY-MM-DD)
        - name: to
          in: query
          schema:
            type: string
          description: End date for range filter (YYYY-MM-DD)
        - name: areaId
          in: query
          schema:
            type: integer
        - name: assignedToId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Issue"
    post:
      operationId: createIssue
      tags: [issues]
      summary: Report an issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIssueRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}/resolve:
    post:
      operationId: resolveIssue
      tags: [issues]
      summary: Resolve an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}/images:
    patch:
      operationId: updateIssueImages
      tags: [issues]
      summary: Attach before/after image paths to an issue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIssueImagesRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}/assign:
    patch:
      operationId: assignIssue
      tags: [issues]
      summary: Assign an issue to a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignIssueRequest"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
  /issues/{id}/complete:
    patch:
      operationId: completeIssue
      tags: [issues]
      summary: Mark an issue complete with notes (staff)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteIssueRequest"
      responses:
        "200":
          description: Completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
...
components:
  securitySchemes:
    InternalCronSecret:
      type: apiKey
      in: header
      name: X-Internal-Cron-Secret
    SendGridInboundSecret:
      type: apiKey
      in: header
      name: X-SendGrid-Inbound-Secret
  schemas:
...
    Issue:
      type: object
      properties:
        id:
          type: integer
        areaId:
          type: integer
        areaName:
          type: string
        reportedById:
          type: integer
        reportedByName:
          type: string
        reportedByActive:
          type: boolean
          description: Whether the staff member who reported the issue is still active.
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        assignedToActive:
          type: boolean
          nullable: true
          description: Whether the assigned staff member is still active. Null when no one is assigned.
        issueDate:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        resolved:
          type: boolean
        resolvedAt:
          type: string
          nullable: true
        completionNotes:
          type: string
          nullable: true
        beforeImagePath:
          type: string
          nullable: true
        afterImagePath:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - areaId
        - areaName
        - reportedById
        - reportedByName
        - reportedByActive
        - issueDate
        - description
        - severity
        - resolved
        - createdAt
    CreateIssueRequest:
      type: object
      properties:
        areaId:
          type: integer
        reportedById:
          type: integer
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        beforeImagePath:
          type: string
          nullable: true
      required:
        - areaId
        - reportedById
        - description
        - severity
    AssignIssueRequest:
      type: object
      properties:
        assignedToId:
          type: integer
          nullable: true
        assignedById:
          type: integer
      required:
        - assignedById
    CompleteIssueRequest:
      type: object
      properties:
        completionNotes:
          type: string
          nullable: true
        completedById:
          type: integer
        afterImagePath:
          type: string
          nullable: true
      required:
        - completedById
    UpdateIssueImagesRequest:
      type: object
      properties:
        beforeImagePath:
          type: string
          nullable: true
        afterImagePath:
          type: string
          nullable: true
...